Client side rendering is done
That took a lot more than I expected, because you can’t reliably inject
new <script> tags. So there’s some more scaffolding in place to link
queries and the API model object, but the end result is the same — the
entire presentation can be written in one place. It’s just two tags - a
template, and a script that executes on load to register the data
handler for the class of items, not the individual instance.