Allow for direct linking to transforms
It’s a little ugly-ish, but here’s what it does:
- set the graphite base URL in the site-header snippet expanded by
flask. That eliminates the ugliness of attempting to store those
parameters. We can remove them from the rendering API call too.
- flask was handling the transform route anyway so those URLs wouldn’t
break in server-side handling, so we can just handle applying the
transform from server side for direct links