Hack jquery.flot.js to get opaque highlights
The only stock way to get opaque highlights in flot is to set the
`color` attribute on each data series. But those are part of the
`Query` object, and shared across multiple presentations, each of which
may have a different color palette.

The only other way to get it is to modify flot itself, which I’ve done
here. I don’t like it, but the 50% transparent highlights were not
sufficiently visible.