Remove transform when entering standard mode
This is a workaround for part of issue #450 - remove transforms when
entering standard mode. This works around what happens when you go
standard -> transform -> display -> standard, which would get stuck
with the transform applied.

The actual fix to return to the previous state (transform applied) is a
better mode stack.