Isolate API calls in a client class
Start of a proper API client to wrap API calls to the tessera server in
a typed, Promise-based class. Most major use of the API has been
converted to the client, but not quite all yet
- Includes a few minor API changes. Instead of `GET
/api/dashboard/<id>?rendering=true`, use `GET
/api/dashboard/<id>/for-rendering`, which now returns a tuple of the
dashboard and prefs, instead of sticking the prefs in the dashboard
- Removed `/api/config`. All settings are returned by
`/api/preferences` now.