Incorporate tessera_client module
Once upon a time, the API client code was split into a separate module,
for re-use without downloading the larger server module.

That never really panned out to be useful, because the front-end code
is the only full implementation of the data model. The client can still
be used separately from the `tessera` pip module if needed.